Move eye_movement utils into separate file
