+++ pupil_src/shared_modules/eye_movement_detector.py
@@ -48,11 +48,6 @@ from pyglui.cygl.utils import RGBA, draw_circle
 from pyglui.pyfontstash import fontstash
 
 
-logger = logging.getLogger(__name__)
-
-EYE_MOVEMENT_EVENT_KEY = 'eye_movement_segments'
-
-
 class Time_Range():
     def __init__(self, start_time: float, end_time: float):
         assert start_time <= end_time
@@ -143,15 +138,6 @@ class Segment_Class(enum.Enum):
             Segment_Class.POST_SACCADIC_OSCILLATIONS: blue,
             Segment_Class.SMOOTH_PURSUIT: purple,
         }.get(self, grey)
-
-
-Gaze_Data = typing.Iterable[fm.Serialized_Dict]
-
-
-Gaze_Time = typing.Iterable[float]
-
-
-MsgPack_Serialized_Segment = typing.Type[bytes]
 
 
 
@@ -539,42 +525,6 @@ class Classified_Segment_Factory:
         id = self._segment_id
         self._segment_id += 1
         return id
-
-
-def can_use_3d_gaze_mapping(gaze_data) -> bool:
-    return "gaze_normal_3d" in gaze_data[0] or "gaze_normals_3d" in gaze_data[0]
-
-
-def clean_3d_data(gaze_points_3d: np.ndarray) -> np.ndarray:
-    # Sometimes it is possible that the predicted gaze is behind the camera which is physically impossible.
-    gaze_points_3d[gaze_points_3d[:, 2] < 0] *= -1.0
-    return gaze_points_3d
-
-
-def np_denormalize(points_2d, frame_size):
-    width, height = frame_size
-    points_2d[:, 0] *= width
-    points_2d[:, 1] = (1.0 - points_2d[:, 1]) * height
-    return points_2d
-
-
-def preprocess_eye_movement_data(capture, gaze_data, use_pupil: bool):
-
-    if use_pupil:
-        assert can_use_3d_gaze_mapping(gaze_data)
-        gaze_points_3d = [gp["gaze_point_3d"] for gp in gaze_data]
-        gaze_points_3d = np.array(gaze_points_3d, dtype=np.float32)
-        gaze_points_3d = clean_3d_data(gaze_points_3d)
-    else:
-        gaze_points_2d = np.array([gp["norm_pos"] for gp in gaze_data])
-        gaze_points_2d = np_denormalize(gaze_points_2d, capture.frame_size)
-        gaze_points_3d = capture.intrinsics.unprojectPoints(gaze_points_2d)
-
-    x, y, z = gaze_points_3d.T
-    r, theta, psi = methods.cart_to_spherical([x, y, z])
-
-    nslr_data = np.column_stack([theta, psi])
-    return nslr_data
 
 
 Eye_Movement_Generator_Yield = typing.Tuple[
